+++ src/modules/translation/translation.types.ts
@@ -1,4 +1,12 @@
+import { MealRecipeSections } from '../meal/meal.types';
+
 export interface TranslatedIngredient {
     text: string;
     imageUrl: string;
+}
+
+export interface TranslatedDetailedMeal {
+    description: string;
+    ingredients: TranslatedIngredient[];
+    recipe: MealRecipeSections;
 }