+++ test/image.e2e-spec.ts
@@ -2,11 +2,11 @@ import { INestApplication } from '@nestjs/common';
 import { ImageService } from '../src/modules/image/image.service';
 import { Test } from '@nestjs/testing';
 import { AppModule } from '../src/app.module';
-import * as cookieParser from 'cookie-parser';
+import cookieParser from 'cookie-parser';
 import { MealService } from '../src/modules/meal/meal.service';
 import { JwtManagerService } from '../src/modules/jwt-manager/jwt-manager.service';
-import * as request from 'supertest';
-import * as path from 'path';
+import request from 'supertest';
+import path from 'path';
 import { LoggerService } from '../src/modules/logger/logger.service';
 import { RedisService } from '../src/modules/redis/redis.service';